Organize the components folder
