+++ src/components/ScriptSettings/ScriptSettings.vue
