+++ src/components/status/SimpleDeviceStatus.vue
