+++ src/components/status/StatusColumn.vue
