+++ src/components/status/StatusRow.vue
